pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "logzai-otlp"
version = "1.0.7"
description = "OpenTelemetry logging client for LogzAI"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "LogzAI" }]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
  "opentelemetry-sdk>=1.27.0",
  "opentelemetry-exporter-otlp>=1.27.0",
  "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://logzai.com"

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "build>=1.3.0",
    "twine>=6.2.0",
]
examples = [
    "fastapi>=0.127.0",
    "langchain>=0.3.27",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.0",
    "logfire>=4.15.1",
    "pydantic>=2.12.5",
    "pydantic-ai>=0.8.1",
    "python-dotenv>=1.0.0",
]
